# api/src/openapi.yml
openapi: 3.0.3
info:
  title: Offline-First Chat Application API
  description: REST API for an offline-first chat application with event-based synchronization and configurable conflict resolution
  version: 1.0.0
servers:
  - url: https://api.chatapp.com/v1
    description: Production server
paths:
...
  /sync/membershipEvents:
    post:
      summary: Synchronize membership events (offline-first sync)
      tags: [Synchronization]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                events:
                  type: array
                  items:
                    $ref: '#/components/schemas/PendingMembershipEvent'
                lastSyncTimestamp:
                  type: string
                  format: date-time
                  description: Client's last successful sync timestamp
              required: [events, last_sync_timestamp]
      responses:
        '200':
          description: Sync completed with conflict resolution
          content:
            application/json:
              schema:
                type: object
                properties:
                  conflictsResolved:
                    type: array
                    items:
                      $ref: '#/components/schemas/ResolvedConflict'
                  serverEvents:
                    type: array
                    items:
                      $ref: '#/components/schemas/MembershipEvent'
                  syncTimestamp:
                    type: string
                    format: date-time
                  acceptedEvents:
                    type: array
                    items:
                      type: object
                      properties:
                        localId:
                          type: string
                        serverId:
                          type: string
                          format: uuid
...
  /sync/messages:
    post:
      summary: Synchronize messages (offline-first sync)
      tags: [Synchronization]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                messages:
                  type: array
                  items:
                    $ref: '#/components/schemas/PendingMessage'
                lastSyncTimestamp:
                  type: string
                  format: date-time
              required: [messages, last_sync_timestamp]
      responses:
        '200':
          description: Message sync completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  conflictsResolved:
                    type: array
                    items:
                      $ref: '#/components/schemas/ResolvedConflict'
                  serverMessages:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
                  syncTimestamp:
                    type: string
                    format: date-time
                  acceptedMessages:
                    type: array
                    items:
                      type: object
                      properties:
                        localId:
                          type: string
                        serverId:
                          type: string
                          format: uuid

  # Admin Endpoints
...
  /auth/register:
    post:
      summary: Create a new user account
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  minLength: 3
                  maxLength: 50
                displayName:
                  type: string
                  maxLength: 100
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
              required: [username, display_name, email, password]
      responses:
        '201':
          description: Account created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
        '409':
          description: Username or email already exists
  /auth/login:
    post:
      summary: Login with existing account
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              required: [username, password]
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    AdminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin-level JWT token required
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        displayName:
          type: string
        isAdmin:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    Message:
      type: object
      properties:
        id:
          type: string
          format: uuid
        groupId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        username:
          type: string
        content:
          type: string
        messageType:
          type: string
          enum: [text, system]
        createdAt:
          type: string
          format: date-time
        syncStatus:
          type: string
          enum: [synced, pending, failed]
    PendingMessage:
      type: object
      properties:
        localId:
          type: string
        groupId:
          type: string
          format: uuid
        content:
          type: string
        messageType:
          type: string
          enum: [text, system]
          default: text
        createdAt:
          type: string
          format: date-time
          description: Client timestamp when message was created
      required: [local_id, group_id, content, created_at]
    MembershipEvent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        groupId:
          type: string
          format: uuid
        action:
          type: string
          enum: [JOIN, LEAVE, REMOVE]
        performedBy:
          type: string
          format: uuid
          description: User who performed the action (for REMOVE actions)
        timestamp:
          type: string
          format: date-time
    PendingMembershipEvent:
      type: object
      properties:
        localId:
          type: string
        groupId:
          type: string
          format: uuid
        action:
          type: string
          enum: [JOIN, LEAVE]
        timestamp:
          type: string
          format: date-time
          description: Client timestamp when action was performed
      required: [local_id, group_id, action, timestamp]
    ResolvedConflict:
      type: object
      properties:
        conflictId:
          type: string
          format: uuid
        conflictType:
          type: string
          enum:
            [
              SIMULTANEOUS_LEAVE_ADD,
              DUPLICATE_JOIN,
              DUPLICATE_LEAVE,
              SEND_TO_LEFT_GROUP,
              SEND_TO_DELETED_GROUP,
              DUPLICATE_MESSAGE,
              MESSAGE_ORDER_CONFLICT,
            ]
        resolutionApplied:
          type: string
          description: Name of the resolution strategy that was applied
        userMessage:
          type: string
          description: Human-readable message to show to the user
        resolvedState:
          type: object
          description: The final state after conflict resolution
          additionalProperties: true
...